requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
alembic
passlib[bcrypt]
python-jose
pydantic
